Rename package 'collection.items' to 'collection.assets'
